Module pallet_amm::pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Structs

Instance1
Instance1 to be used for instantiable pallet define with pallet macro.

Instance2
Instance2 to be used for instantiable pallet define with pallet macro.

Instance3
Instance3 to be used for instantiable pallet define with pallet macro.

Instance4
Instance4 to be used for instantiable pallet define with pallet macro.

Instance5
Instance5 to be used for instantiable pallet define with pallet macro.

Instance6
Instance6 to be used for instantiable pallet define with pallet macro.

Instance7
Instance7 to be used for instantiable pallet define with pallet macro.

Instance8
Instance8 to be used for instantiable pallet define with pallet macro.

Instance9
Instance9 to be used for instantiable pallet define with pallet macro.

Instance10
Instance10 to be used for instantiable pallet define with pallet macro.

Instance11
Instance11 to be used for instantiable pallet define with pallet macro.

Instance12
Instance12 to be used for instantiable pallet define with pallet macro.

Instance13
Instance13 to be used for instantiable pallet define with pallet macro.

Instance14
Instance14 to be used for instantiable pallet define with pallet macro.

Instance15
Instance15 to be used for instantiable pallet define with pallet macro.

Instance16
Instance16 to be used for instantiable pallet define with pallet macro.

Pallet
The pallet implementing the on-chain logic.

_GeneratedPrefixForStorageExchangeRate

_GeneratedPrefixForStorageLiquidityProviders

_GeneratedPrefixForStoragePools

Enums

Call
Contains one variant per dispatchable that can be called by an extrinsic.

Error
Custom dispatch errors of this pallet.

Event
The event emitted by this pallet.

Traits

Config
Configuration trait of this pallet.

Type Definitions

ExchangeRate
The exchange rate from the underlying to the internal collateral

LiquidityProviders
Accounts that deposits and withdraw assets in one or more pools

Module	Deprecated
Type alias to Pallet, to be used by construct_runtime.

Pools
A bag of liquidity composed by two different assets